requirements.txt
asgiref==3.5.2
azure-common==1.1.28
azure-core==1.26.2
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==3.2.16
idna==3.4
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.6.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
